add delete face

+++ controllers/UserController.js
@@ -122,7 +122,6 @@ export const setPin = async (req, res, next) => {
 export const addFace = async (req, res, next) => {
     const {userID, name, images} = req.body
 
-    console.log('alo')
     FaceID.create({ userID, name, images })
         .then((image) =>    
             res.status(201).json({
@@ -136,10 +135,22 @@ export const addFace = async (req, res, next) => {
 
 //[POST]/api/users/deleteface
 export const deleteFace = async (req, res, next) => {
-    const { face_id } = req.body
+    const { face_id, user_id } = req.body
+    if (!face_id) {
+        res.status(400)
+        next(new Error('Bạn phải điền các thông tin cần thiết'))
+    }
     
     FaceID.deleteOne({_id: face_id})
-        .then((e) => res.status(200).json({ message: "Xoá thành công" }))
+        .then((e) => {
+            FaceID.find({userID: user_id})
+            .then(face => 
+                {
+                    res.status(201).json(face)    
+                }
+            )
+            .catch(next)
+        })
         .catch(next)
 }
